New get data by id method, new update method, and new delete method introduced.

+++ SolarWatch/SolarWatch/Services/SWServices/SWRepository.cs
@@ -1,3 +1,4 @@
+using Microsoft.EntityFrameworkCore;
 using SolarWatch.Data;
 using SolarWatch.Model;
 using SolarWatch.Utilities;
@@ -8,40 +9,59 @@ public class SWRepository : ISWRepository
 {
     private readonly DataContext _context;
     private readonly INormalizeCityName _normalizeCityName;
-    
+
     public SWRepository(DataContext context, INormalizeCityName normalizeCityName)
     {
         _context = context;
         _normalizeCityName = normalizeCityName;
     }
-    
+
     public IEnumerable<SWData> GetAllSWDatas()
     {
         return _context.SolarWatchDatas.ToList();
     }
-    
+
     public SWData GetSWData(string city, DateOnly date)
     {
         city = _normalizeCityName.Normalize(city);
         return _context.SolarWatchDatas.FirstOrDefault(c => c.City == city && c.Date == date);
     }
-    
+
+    public async Task<SWData> GetSWDataById(int id)
+    {
+        return await _context.SolarWatchDatas.FirstOrDefaultAsync(c => c.Id == id);
+    }
+
     public void AddSWData(SWData swData)
     {
         _context.SolarWatchDatas.Add(swData);
         _context.SaveChanges();
     }
-    
-    public void UpdateSWData(SWData swData)
+
+    public async Task UpdateSWData(SWData swData)
     {
-        _context.SolarWatchDatas.Update(swData);
-        _context.SaveChanges();
+        var swDataFromDb = await _context.SolarWatchDatas.FirstOrDefaultAsync(c => c.Id == swData.Id);
+        if (swDataFromDb != null)
+        {
+            swDataFromDb.City = swData.City;
+            swDataFromDb.Date = swData.Date;
+            swDataFromDb.Sunrise = swData.Sunrise;
+            swDataFromDb.Sunset = swData.Sunset;
+
+            _context.SolarWatchDatas.Update(swDataFromDb);
+            await _context.SaveChangesAsync();
+        }
     }
-    
-    public void DeleteSWData(SWData swData)
+
+    public void DeleteSWData(int id)
     {
-        _context.SolarWatchDatas.Remove(swData);
-        _context.SaveChanges();
+        var swData = _context.SolarWatchDatas.FirstOrDefault(c => c.Id == id);
+        if (swData != null)
+        {
+            _context.SolarWatchDatas.Remove(swData);
+            _context.SaveChanges();
+        }
+
     }
-    
+
 }
